src: fix Qt 6.11 + AppleClang 17 warnings (caught by macOS release build)
The local `make release` build (which mirrors CI's release.yml with
-Werror) flagged a handful of warnings that GCC + Qt 6.8 in CI doesn't
trip but AppleClang 17 + Qt 6.11 does:

- Application.cpp: drop unused `[this]` capture; QApplication::
  activeWindow() is static.

- ImageAdapter.cpp: QImage::mirrored() is deprecated in Qt 6.13;
  Qt 6.9 added flipped(Qt::Orientations). Use the latter under a
  QT_VERSION_CHECK guard until the minimum Qt moves to 6.9.

- AnnotationOverlay.cpp: float→qreal promotion in std::clamp call
  was implicit; make the cast explicit.

- MainWindow.cpp: drop unused `[this]` capture in the SAM/OCR
  download wireSignals lambda.

- ModelManagerDialog.cpp: qsizetype→int narrowing for setRowCount /
  loop bound; cast at the boundary.

- test_uat_foundations.cpp / test_uat_search_and_markup.cpp: same
  family of qsizetype→int and int→size_t conversion warnings in
  the UAT PDF-fixture builders.

None of these change runtime behaviour; they make the source clean
under stricter warning configs.

+++ tests/uat/test_uat_search_and_markup.cpp
@@ -106,7 +106,7 @@ QString writePdfWithOutline(const QString& path,
         return r;
     };
 
-    const int pageCount = titles.size();
+    const int pageCount = static_cast<int>(titles.size());
 
     // Build /Pages root first so each page's /Parent can point at it.
     QPDFObjectHandle pagesDict = QPDFObjectHandle::newDictionary();
@@ -147,7 +147,7 @@ QString writePdfWithOutline(const QString& path,
         item.replaceKey("/Parent", outlinesObj);
         // Destination: jump to top of page at current zoom.
         QPDFObjectHandle dest = QPDFObjectHandle::newArray();
-        dest.appendItem(pageObjs[i]);
+        dest.appendItem(pageObjs[static_cast<size_t>(i)]);
         dest.appendItem(QPDFObjectHandle::newName("/XYZ"));
         dest.appendItem(QPDFObjectHandle::newInteger(0));
         dest.appendItem(QPDFObjectHandle::newInteger(792));
@@ -155,9 +155,9 @@ QString writePdfWithOutline(const QString& path,
         item.replaceKey("/Dest", dest);
         itemObjs.push_back(pdf.makeIndirectObject(item));
     }
-    for (int i = 0; i < int(itemObjs.size()); ++i) {
+    for (size_t i = 0; i < itemObjs.size(); ++i) {
         if (i > 0) itemObjs[i].replaceKey("/Prev", itemObjs[i - 1]);
-        if (i + 1 < int(itemObjs.size())) {
+        if (i + 1 < itemObjs.size()) {
             itemObjs[i].replaceKey("/Next", itemObjs[i + 1]);
         }
     }
